Create README.md
Our group is thinking of creating a “choose your own path” game. We will create a map, and a user will be able to choose where he or she wants to go on this map. Once the user is at their desired location, they will be given two options. They choose which ever option sounds more interesting to them, and they will move to a new location on the map. The process will repeat until the user chooses “the wrong path” or chooses to quit. We want to use OpenGL to create an image of the map.