Contributions
The first change was to the play/exit prompt which made it easier to
follow and more navigatable.

Second change was to automatically turn every character lower case. I
removed the cout statement instructing users to only input lowercase
characters.

Final change was if the user entered an input with either a digit or
whitespace they would be told that their input was invalid, but would
not face a penalty for doing so. They would then press enter and go back
to where they were. (Made code easy to change if owners want to penalize
user for bad input)